# documentation/openapi.yaml
openapi: 3.1.0
info:
  title: Hunting Mr. X Wellington Edition API
  description: |
    REST API for the Hunting Mr. X: Wellington Edition multiplayer game backend.

    ## Real-time updates (WebSocket / STOMP)

    The backend also exposes a STOMP-over-SockJS endpoint at `/ws`.
    Clients connect using `@stomp/stompjs` + `sockjs-client`.

    **Lobby topic** — subscribe after joining or creating a game:
    ```
    /topic/games/{gameId}
    ```
    Published after `join`, `kick`, and `leave` (lobby phase only).

    **Per-player topic** — subscribe on game board mount using the player's own ID:
    ```
    /topic/games/{gameId}/players/{playerId}
    ```
    Published after every in-progress state change. Role-filtered: Mr X's
    `nodeId` is hidden from detectives except on reveal rounds (2, 8, 13, 18, 24).

    **Valid moves topic** — subscribe alongside the per-player topic:
    ```
    /topic/games/{gameId}/players/{playerId}/valid-moves
    ```
    Published at the start of a player's turn. Each entry lists a reachable
    `nodeId` and the `ticketOptions` the player may use to reach it.

    **Kick detection** — a client should check on each push whether
    their own `playerId` is still present in `players`. If it is not
    and `phase` is still `LOBBY`, the client was kicked.
  version: 0.1.11
servers:
  - url: http://localhost:8080
    description: Local development server

paths:

  /api/map:
    get:
      summary: Get the active game map
      description: |
        Returns the pre-computed Wellington graph currently configured via
        `game.map-file` in `application.properties`. The backend reads that
        file from the classpath and streams it as JSON — so swapping the
        property (e.g. `test-map.json` → `map.json`) changes what both the
        game engine and the frontend receive without any frontend change.

        The frontend fetches this once on game board mount.
      operationId: getMap
      responses:
        '200':
          description: Map data loaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MapData'
        '500':
          description: Configured map file not found in classpath

  /api/games/create:
    post:
      summary: Create a new game
      description: Creates a new game session. The caller becomes the host. Returns the host's player ID and the initial game state.
      operationId: createGame
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGameRequest'
            example:
              hostName: Alice
              maxPlayers: 4
      responses:
        '201':
          description: Game created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateGameResponse'
        '400':
          description: Validation failure — host name blank or maxPlayers out of range (2–6)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Game not created

  /api/games/join:
    post:
      summary: Join an existing game
      description: |
        Joins a LOBBY-phase game identified by its join code. Returns the joining
        player's ID and the updated game state. Also broadcasts the updated
        `GameState` to `/topic/games/{gameId}` so all lobby subscribers see
        the new player immediately.
      operationId: joinGame
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinGameRequest'
            example:
              joinCode: WXYZ12
              playerName: Bob
      responses:
        '200':
          description: Joined successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JoinGameResponse'
        '400':
          description: Validation failure — join code or player name blank
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Join code is required
        '404':
          description: No game with that join code exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Game not found
        '409':
          description: Game is not accepting new players — either not in LOBBY phase or already full
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                wrongPhase:
                  summary: Game already started
                  value:
                    error: Game is not in the lobby phase
                full:
                  summary: Game is full
                  value:
                    error: Game is full

  /api/games/{id}:
    get:
      summary: Get game state
      description: |
        Returns the current state of a game. When `playerId` is supplied the
        response is role-filtered: Mr X's `nodeId` is hidden from detectives on
        non-reveal rounds, and `mrXLog` entries omit the node on non-reveal legs.
        Omit `playerId` for an unfiltered view (lobby use, spectators).
      operationId: getGame
      parameters:
        - $ref: '#/components/parameters/GameId'
        - name: playerId
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: If supplied, response is role-filtered for this player
      responses:
        '200':
          description: Current game state (role-filtered when playerId provided)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameState'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Game not found

  /api/games/{id}/start:
    post:
      summary: Start the game
      description: |
        Transitions the game from `LOBBY` to `IN_PROGRESS`. Only the host may
        call this. Requires at least 2 players.

        One player is randomly assigned `MR_X` (unlimited transport tickets,
        2 DOUBLE, N BLACK/"Invisible" where N = detective count); the rest
        become `DETECTIVE` (finite ticket counts from server config).

        Broadcasts the updated `GameState` to `/topic/games/{gameId}` — all
        lobby subscribers detect `phase: IN_PROGRESS` and navigate to the game
        board.
      operationId: startGame
      parameters:
        - $ref: '#/components/parameters/GameId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartGameRequest'
            example:
              playerId: 550e8400-e29b-41d4-a716-446655440000
      responses:
        '200':
          description: Game started; returns updated state with roles and tickets assigned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameState'
        '400':
          description: Not enough players (minimum 2 required)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Need at least 2 players to start
        '403':
          description: Caller is not the host
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Only the host can start the game
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Game not found
        '409':
          description: Game is not in LOBBY phase
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Game is not in the lobby phase

  /api/games/{id}/valid-moves:
    get:
      summary: Get valid moves for the current player
      description: |
        Returns the list of nodes the specified player may move to on their
        current turn, together with which ticket types they may use on each edge.

        The server checks adjacency, ticket availability, and blocking (detective
        positions block Mr X). For a DOUBLE first leg, any adjacent non-blocked
        node is valid regardless of ticket type. For the second leg and normal
        turns, only edges whose transport mode the player has a ticket for are
        included.

        The server also pushes this data to
        `/topic/games/{gameId}/players/{playerId}/valid-moves` at the start of
        each turn; this REST endpoint is a fallback for clients that missed the push.
      operationId: getValidMoves
      parameters:
        - $ref: '#/components/parameters/GameId'
        - name: playerId
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the player requesting their valid moves
      responses:
        '200':
          description: Valid moves computed successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ValidMove'
        '400':
          description: It is not this player's turn
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Player is not in this game
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Game or player not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/games/{id}/moves:
    post:
      summary: Submit a move
      description: |
        The active player submits their chosen destination node and the ticket
        they wish to use. The server validates the move (adjacency, ticket
        availability, turn ownership) and advances the game state.

        **Mr X DOUBLE move:** submit with `ticket: "DOUBLE"` to start a double
        move. The server sets `mrXDoubleMovePending: true` and pushes a new set
        of valid moves for the second leg. Submit again with the real transport
        ticket to complete the double move.

        **Catch check:** after each detective move the server tests whether the
        detective landed on Mr X's node. If so, `phase` → `ENDED`,
        `winner: "DETECTIVES"`.

        **Round end:** after all detectives have moved, the round counter
        increments. At round 24 Mr X wins.

        After a successful move the server broadcasts the updated
        `GameState` to all per-player topics and pushes new valid moves
        to the next active player.
      operationId: submitMove
      parameters:
        - $ref: '#/components/parameters/GameId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveRequest'
            example:
              playerId: 550e8400-e29b-41d4-a716-446655440000
              toNodeId: 42
              ticket: BUS
      responses:
        '200':
          description: Move accepted; returns the updated game state (role-filtered for the submitting player)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameState'
        '400':
          description: Move is invalid (wrong turn, invalid node, insufficient tickets)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: It is not this player's turn
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Game or player not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/games/{id}/players/{targetPlayerId}:
    delete:
      summary: Remove a player from the game (leave or kick)
      description: |
        Removes a player from the game. Two authorised callers:

        - **The player themselves** (`targetPlayerId` matches their own ID) —
          interpreted as a voluntary leave.
        - **The host** (host sends their `requesterId` in the request body with
          a different `targetPlayerId`) — interpreted as a kick. Only valid
          during the LOBBY phase.

        **Leave behaviour by phase:**
        - Lobby, host leaves → game session deleted entirely; no broadcast.
        - Lobby, non-host leaves → player removed; broadcasts updated state.
        - In-progress, Mr X leaves → `phase: ENDED`, `abortReason` set;
          broadcasts final state.
        - In-progress, detective leaves → player removed; broadcasts updated
          state.

        **Kick behaviour:**
        - Only valid during `LOBBY`. Host cannot kick themselves.
        - Broadcasts updated `GameState`; kicked client detects removal by
          checking that their `playerId` is absent from `players`.
      operationId: removePlayer
      parameters:
        - $ref: '#/components/parameters/GameId'
        - name: targetPlayerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID of the player to remove
      requestBody:
        required: false
        description: |
          Required only when the host is kicking another player.
          Omit (or set `requesterId` equal to `targetPlayerId`) for a
          voluntary leave.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemovePlayerRequest'
            example:
              requesterId: 550e8400-e29b-41d4-a716-446655440000
      responses:
        '204':
          description: Player removed successfully (no body)
        '400':
          description: Host attempted to kick themselves
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Host cannot kick themselves
        '403':
          description: Caller is not the host (kick attempt only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Only the host can kick players
        '404':
          description: Game or player not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Game or player not found
        '409':
          description: Kick attempted outside of LOBBY phase
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: Players can only be kicked during the lobby

# ---------------------------------------------------------------------------
# WebSocket push messages (STOMP over SockJS at /ws)
#
# OpenAPI 3.1.0 has no native WebSocket support. The `webhooks` section is
# used here to formally describe server-initiated STOMP messages — each entry
# is one topic the client can subscribe to. The payload schema is the same
# GameState used by the REST endpoints.
# ---------------------------------------------------------------------------
webhooks:

  lobbyStatePush:
    post:
      summary: Lobby-phase game state broadcast
      description: |
        **STOMP topic:** `/topic/games/{gameId}`

        The server publishes a `GameState` to this topic during the lobby
        phase after join, kick, and leave operations. Subscribe immediately
        after creating or joining a game. Unsubscribe once `phase` transitions
        to `IN_PROGRESS` (switch to the per-player topic instead).

        **Triggered by:**
        | REST call | What changes |
        |---|---|
        | `POST /api/games/join` | New player added to `players` |
        | `POST /api/games/{id}/start` | `phase` → `IN_PROGRESS`; per-player topics take over |
        | `DELETE /api/games/{id}/players/{targetPlayerId}` (kick) | Target player removed from `players` |
        | `DELETE /api/games/{id}/players/{targetPlayerId}` (leave, lobby) | Player removed |

        **Client-side reactions:**
        - `phase === 'IN_PROGRESS'` → navigate to game board; subscribe to per-player topic
        - Own `playerId` absent from `players` while `phase === 'LOBBY'` → kicked; show banner

        **Connection:** `new SockJS('/ws')` → STOMP `Client` → subscribe on `onConnect`.
      requestBody:
        description: The updated game state pushed by the server.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameState'
      responses:
        '200':
          description: Client acknowledges the message (STOMP ACK).

  perPlayerGameStatePush:
    post:
      summary: Per-player in-game state broadcast
      description: |
        **STOMP topic:** `/topic/games/{gameId}/players/{playerId}`

        The server publishes a role-filtered `GameState` to each player's
        own topic after every in-progress state change (move submission, turn
        advancement, win/abort). Subscribe on game board mount using the
        viewer's own `playerId`. The payload hides Mr X's `nodeId` from
        detectives on non-reveal rounds.

        **Triggered by:**
        | Event | What changes |
        |---|---|
        | `POST /api/games/{id}/moves` | Player position updated; turn advanced |
        | Mr X caught | `phase` → `ENDED`, `winner: "DETECTIVES"` |
        | Round 24 complete | `phase` → `ENDED`, `winner: "MR_X"` |
        | Player leaves in-progress | `phase` → `ENDED`, `abortReason` set |

        **Client-side reactions:**
        - `phase === 'ENDED'` → navigate to `/game/{id}/end`
        - `mrXDoubleMovePending === true` → show "second leg" banner; await new valid-moves push
        - Update map markers and ticket counts from the new state

        **Reveal rounds:** 2, 8, 13, 18, 24. On these rounds the final
        `mrXLog` entry has a non-null `nodeId` and Mr X's position is
        shown in detective-view responses.
      requestBody:
        description: Role-filtered game state for the subscribing player.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameState'
      responses:
        '200':
          description: Client acknowledges the message (STOMP ACK).

  validMovesPush:
    post:
      summary: Valid moves push at turn start
      description: |
        **STOMP topic:** `/topic/games/{gameId}/players/{playerId}/valid-moves`

        The server publishes a bare array of `ValidMove` to the active player
        at the start of their turn (immediately after `startGame` for Mr X;
        after each move submission for the next player). Subscribe on game
        board mount alongside the per-player state topic.

        The client uses this array to highlight reachable nodes on the
        map and populate the ticket-picker. Call
        `GET /api/games/{id}/valid-moves?playerId={pid}` if the push is missed.

        **Triggered by:**
        - Game start (pushes to Mr X)
        - Each accepted move (pushes to the next active player)
        - DOUBLE first-leg accepted (pushes second-leg moves to Mr X)
      requestBody:
        description: Valid destinations and ticket options for the active player's turn.
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/ValidMove'
      responses:
        '200':
          description: Client acknowledges the message (STOMP ACK).

components:

  parameters:
    GameId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: UUID of the game session

  schemas:

    CreateGameRequest:
      type: object
      required: [hostName, maxPlayers]
      properties:
        hostName:
          type: string
          minLength: 1
          description: Display name of the host player
        maxPlayers:
          type: integer
          minimum: 2
          maximum: 6
          description: Total player cap (including Mr X)

    CreateGameResponse:
      type: object
      required: [playerId, gameState]
      properties:
        playerId:
          type: string
          format: uuid
          description: UUID assigned to the host; must be stored client-side
        gameState:
          $ref: '#/components/schemas/GameState'

    JoinGameRequest:
      type: object
      required: [joinCode, playerName]
      properties:
        joinCode:
          type: string
          minLength: 6
          maxLength: 6
          pattern: '^[A-Z0-9]{6}$'
          description: 6-character uppercase join code shown in the lobby
        playerName:
          type: string
          minLength: 1
          description: Display name chosen by the joining player

    JoinGameResponse:
      type: object
      required: [playerId, gameState]
      properties:
        playerId:
          type: string
          format: uuid
          description: UUID assigned to the joining player; must be stored client-side
        gameState:
          $ref: '#/components/schemas/GameState'

    StartGameRequest:
      type: object
      required: [playerId]
      properties:
        playerId:
          type: string
          format: uuid
          description: UUID of the player requesting the start (must be the host)

    RemovePlayerRequest:
      type: object
      properties:
        requesterId:
          type: string
          format: uuid
          description: |
            UUID of the player making the request. Required when the host is
            kicking another player. Omit for a voluntary leave.

    MoveRequest:
      type: object
      required: [playerId, toNodeId, ticket]
      properties:
        playerId:
          type: string
          format: uuid
          description: UUID of the player submitting the move
        toNodeId:
          type: integer
          description: Destination node ID
        ticket:
          $ref: '#/components/schemas/TicketType'
          description: |
            Ticket to consume. Use `DOUBLE` to initiate a double move (Mr X only).
            On the second leg of a double move, provide the real transport ticket.

    ValidMove:
      type: object
      required: [nodeId, ticketOptions]
      properties:
        nodeId:
          type: integer
          description: Reachable destination node ID
        ticketOptions:
          type: array
          items:
            type: string
          description: |
            Ticket types the player may use to reach this node.
            For Mr X may include `DOUBLE` (on normal turns) and `BLACK`
            (shown to players as "Invisible").
            For detectives, only the transport modes of the connecting edge.

    MrXLogEntryView:
      type: object
      required: [round, leg, ticketUsed]
      properties:
        round:
          type: integer
          description: Game round this move occurred in
        leg:
          type: integer
          enum: [1, 2]
          description: 1 for a normal move or first leg of a double; 2 for the second leg
        ticketUsed:
          $ref: '#/components/schemas/TicketType'
        nodeId:
          type: integer
          nullable: true
          description: |
            Mr X's destination node. Non-null only on the final leg of a reveal
            round (rounds 2, 8, 13, 18, 24). Null on all other entries.

    GameState:
      type: object
      required: [gameId, joinCode, phase, maxPlayers, players, round]
      properties:
        gameId:
          type: string
          format: uuid
        joinCode:
          type: string
          description: 6-character uppercase room code
          example: WXYZ12
        phase:
          $ref: '#/components/schemas/GamePhase'
        maxPlayers:
          type: integer
          minimum: 2
          maximum: 6
        players:
          type: array
          items:
            $ref: '#/components/schemas/PlayerView'
        round:
          type: integer
          minimum: 0
          description: Current round number (0 while in lobby)
        turnPhase:
          $ref: '#/components/schemas/TurnPhase'
          nullable: true
        currentPlayerId:
          type: string
          format: uuid
          nullable: true
          description: UUID of the player whose turn it is; null outside IN_PROGRESS
        winner:
          type: string
          nullable: true
          description: '"MR_X", "DETECTIVES", or null if game is still running'
        abortReason:
          type: string
          nullable: true
          description: Human-readable reason the game was aborted (e.g. "Mr. X has left the game"); null if not aborted
        mrXLog:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/MrXLogEntryView'
          description: |
            Chronological log of Mr X's moves. Null during lobby phase.
            In detective-view responses, `nodeId` is null except on reveal-round entries.
        mrXDoubleMovePending:
          type: boolean
          description: |
            True when Mr X has submitted a DOUBLE ticket and the second leg is
            still outstanding. When true, the valid-moves push for the second leg
            has already been sent; other players are blocked until the second move
            is submitted.

    PlayerView:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        role:
          $ref: '#/components/schemas/Role'
          nullable: true
          description: Null until game starts
        nodeId:
          type: integer
          nullable: true
          description: Current map node ID; null until game starts. For Mr X, null in detective-view responses on non-reveal rounds.
        tickets:
          type: object
          nullable: true
          description: |
            Ticket counts keyed by TicketType. Null until game starts.
            `-1` means unlimited (Mr X transport tickets only).
            Mr X holds: ESCOOTER(-1), BUS(-1), TRAIN(-1), FERRY(-1), DOUBLE(2), BLACK(N detectives).
            Detectives hold: ESCOOTER, BUS, TRAIN, FERRY with finite counts from server config. No DOUBLE or BLACK.
          additionalProperties:
            type: integer
          example:
            ESCOOTER: -1
            BUS: -1
            TRAIN: -1
            FERRY: -1
            DOUBLE: 2
            BLACK: 3

    GamePhase:
      type: string
      enum: [LOBBY, IN_PROGRESS, PAUSED, ENDED]
      description: |
        - `LOBBY`: waiting for players, game not yet started
        - `IN_PROGRESS`: game running
        - `PAUSED`: a player disconnected; grace period active
        - `ENDED`: game finished (win, loss, or abort)

    TurnPhase:
      type: string
      enum: [MR_X_TURN, DETECTIVE_TURN]
      description: Sub-phase within an IN_PROGRESS round

    Role:
      type: string
      enum: [MR_X, DETECTIVE]

    TicketType:
      type: string
      enum: [ESCOOTER, BUS, TRAIN, FERRY, BLACK, DOUBLE]
      description: |
        `BLACK` is shown to players as the **Invisible** ticket — it lets
        Mr X travel any edge, any mode, without needing a matching transport
        ticket or revealing which mode was actually used.

    MapData:
      type: object
      required: [nodes, edges]
      description: Pre-computed Wellington graph used by the game board map view
      properties:
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/GraphNode'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/GraphEdge'

    GraphNode:
      type: object
      required: [id, lat, lng, label]
      properties:
        id:
          type: integer
          description: Unique node identifier
        lat:
          type: number
          format: double
          description: Latitude in decimal degrees
        lng:
          type: number
          format: double
          description: Longitude in decimal degrees
        label:
          type: string
          description: Human-readable node name

    GraphEdge:
      type: object
      required: [from, to, modes]
      properties:
        from:
          type: integer
          description: ID of the source node
        to:
          type: integer
          description: ID of the destination node
        modes:
          type: array
          items:
            $ref: '#/components/schemas/TicketType'
          description: Transport modes available on this edge
        coordinates:
          type: array
          description: Road-following polyline as [lng, lat] pairs; omitted for straight-line edges
          items:
            type: array
            items:
              type: number
            minItems: 2
            maxItems: 2

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message